--- src/main/java/kr/hhplus/be/server/application/coupon/CouponIssueEvent.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.application.coupon;
-
-import kr.hhplus.be.server.domain.coupon.CouponEntity;
-
-public record CouponIssueEvent(Long userId, Long couponId) {}
