+++ src/main/java/kr/hhplus/be/server/application/order/DataPlatformClient.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.order;
+
+import kr.hhplus.be.server.domain.order.OrderEntity;
+
+public interface DataPlatformClient {
+
+    void sendToDataPlatform(OrderEntity order);
+}
