+++ src/main/java/kr/hhplus/be/server/application/order/OrderCompletedEvent.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.order;
+
+import kr.hhplus.be.server.domain.order.OrderEntity;
+
+public record OrderCompletedEvent(OrderEntity order) {}
