+++ src/main/java/kr/hhplus/be/server/domain/coupon/CouponIssueEvent.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.domain.coupon;
+
+public record CouponIssueEvent(Long userId, Long couponId) {}
