+++ src/main/java/kr/hhplus/be/server/domain/order/OrderService.java
@@ -25,11 +25,16 @@ import kr.hhplus.be.server.domain.user.UserRepository;
 import kr.hhplus.be.server.domain.user.execption.UserErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
+import org.springframework.web.client.RestTemplate;
 
 @Service
 @RequiredArgsConstructor
@@ -44,6 +49,7 @@ public class OrderService {
     private final ConcurrencyService concurrencyService;
     private final ProductService productService;
     private final CouponService couponService;
+    private final RestTemplate restTemplate; // 데이터플렛폼 전송 restTemplate
 
     private OrderEntity currentOrder; // 보상용 저장
 
@@ -136,7 +142,6 @@ public class OrderService {
                 }
             }
         });
-
     }
     @Transactional
     public void expireOrder(OrderEntity order, PaymentStatus status) {
@@ -144,4 +149,22 @@ public class OrderService {
         orderRepository.save(order);
     }
 
+    // 데이터플렛폼전송
+    @Transactional(propagation = Propagation.REQUIRES_NEW) // 트랜잭션 분리
+    public void sendToDataPlatform(OrderEntity order) {
+        String url = "https://mock-dataplatform.com/api/payments"; // 가상 플랫폼 URL
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+
+        HttpEntity<OrderEntity> entity = new HttpEntity<>(order, headers);
+
+        ResponseEntity<String> response = restTemplate.postForEntity(url, entity, String.class);
+
+        if (response.getStatusCode().is2xxSuccessful()) {
+            System.out.println("데이터 플랫폼에 성공적으로 전송됨");
+        } else {
+            System.err.println("전송 실패: " + response.getStatusCode());
+        }
+    }
 }
